refactor(notebook): share execution and output stores
